Refactor tests/data folder
