+++ tests/Data/ClassConstantsMustHaveVisibility/ClassAWithUnscopedConstant.php
@@ -6,9 +6,9 @@
  * file that was distributed with this source code.
  */
 
-namespace Space;
+namespace PHPStanForPrestaShopTests\Data\ClassConstantsMustHaveVisibility;
 
 class A
 {
-    public $id;
+    const FOO = 1;
 }
