+++ tests/Data/ClassConstantsMustHaveVisibility/ClassCWithPrivateConstant.php
@@ -6,7 +6,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Space;
+namespace PHPStanForPrestaShopTests\Data\ClassConstantsMustHaveVisibility;
 
 class C
 {
