+++ tests/Data/UseStrictTypesForNewClasses/DeclareStrictTypeClassC.php
