+++ tests/Data/UseStrictTypesForNewClasses/NoDeclareStrictTypeClassB.php
