+++ tests/Data/UseStrictTypesForNewClasses/NoDeclareStrictTypeClassSpaceA.php
@@ -6,9 +6,9 @@
  * file that was distributed with this source code.
  */
 
-namespace Space;
+namespace PHPStanForPrestaShopTests\Data\UseStrictTypesForNewClasses;
 
-class B
+class A
 {
     public $id;
 }
